+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./components"
+export { default as Mathbox } from "./Mathbox"
